cabin/pyproject.toml
[tool.poetry]
name = "cabin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cabin", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.26.1"
sqlalchemy = "^2.0.6"
uvicorn = "^0.21.0"
asyncpg = "^0.27.0"
aiosqlite = "^0.18.0"
argon2-cffi = "^21.3.0"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^23.1.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "cabin.app:main"
